backend/requirements.txt
alembic==1.8.1
aniso8601==9.0.1
async-timeout==4.0.2
azure-core==1.26.2
azure-identity==1.12.0
azure-storage-blob==12.14.1
billiard==3.6.4.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==39.0.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Migrate==4.0.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.1
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
Mako==1.2.4
MarkupSafe==2.1.1
msal==1.20.0
msal-extensions==1.0.0
msrest==0.7.1
oauthlib==3.2.2
portalocker==2.7.0
prompt-toolkit==3.0.36
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==1.4.45
twilio==7.15.3
typing_extensions==4.4.0
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
zipp==3.10.0
